Use domains to catch our errors. We still can't control all of the errors that might propogate. The following have not worked to catch something deep in a lib like foundation or jquery:     * Using a domain in a script tag of the `index.html`       The context gets switched to node's context so all of the       files that depend on being in the window's context are busted.     * `window.onerror` Effectively did nothing.     * <body onerror=...> Again, nothing.
The problem is that the resources are loaded in a script tag and
there doesn't seem to be a way handle something that throws in there.
So node-webkit will just die whenever ANY error is unhandled.

The only thing that worked was `process.on('unhandledException')`
But, this is highly unrecommended.
It leaves the node stack in an undefined state.

So, for better or worse,
we might just have to deal with outside libs throwing.
Unless another way is found.

The best way forward is to have everything defined in PureScript.
At least we can control what's happening,
and not have to resort to these wonky solutions.