Tried to hide everything in a directory, but it's not working too well.
There seem to be two issues here:

1.  The distinction between a notebook and a directory is conflated
    with needing to provide a pleasant UX.
    That is, the difference currently is
    whether or not a directory has a file named "index.nb" inside.
2.  We cannot actually know what the structure of the filesystem is
    without sending multiple queries for the entire filesystem.

For the first one, we need to rethink how we create/recognize notebooks,
and the difference between a directory and a notebook.
For the second one, it might be solved by the first,
but it'd still be nice to be able to get back the whole filesystem in one go.

Use new scheme for identifying notebooks.

Fix tests.

Fixes #94.