Finally figure out that duplicate record issue.
Mocked up the status.

Handle ST effects.

Logs get thrown down.

Have to make the buffer into a string.

Don't always show the config.

Restructured a bit and added some styling.

Log connection issues.

In the middle of a refactor.

I think I just need to let theocmpiler find all the issues and it should be back to working.

Revert "In the middle of a refactor."

This reverts commit 623a4116f70c0e2b31c990716d510b87fef749e3.

Add more logging and times.

Fixes #164.